+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Version 2024.3
 --------------
 *in development*
 
+* 🔧 Shows: move Discover screen to top-level.
+* 🔧 Shows: display all search results, filter by release year.
 * 🔧 Crew: display directors and writers first, then in order of job title.
 
 #### 2024.3.0 🧪